[SOIN] Extrais la ressource du profil
