+++ src/configurationServeur.ts
@@ -1,0 +1,5 @@
+import {EntrepotProfil} from "./entrepotProfil.js";
+
+export type ConfigurationServeur = {
+    entrepotProfil: EntrepotProfil;
+};
