+++ src/mpa.ts
@@ -1,26 +1,10 @@
 import express from "express";
-import { EntrepotProfil } from "./entrepotProfil.js";
+import { ressourceProfil } from "./ressourceProfil.js";
+import { ConfigurationServeur } from "./configurationServeur.js";
 
-type ConfigurationServeur = {
-  entrepotProfil: EntrepotProfil;
-};
-
-const creeServeur = ({ entrepotProfil }: ConfigurationServeur) => {
+const creeServeur = (configurationServeur: ConfigurationServeur) => {
   const app = express();
-
-  app.get("/profil", (requete, reponse) => {
-    if (!requete.query.email) {
-      reponse.sendStatus(400);
-      return;
-    }
-    const profil = entrepotProfil.parEmail(requete.query.email as string);
-    if (!profil) {
-      reponse.sendStatus(404);
-      return;
-    }
-
-    reponse.send(profil);
-  });
+  app.use("/profil", ressourceProfil(configurationServeur));
 
   return app;
 };
